refactor: split metrics module into multiple files
