+++ filter/metrics/filter.go
@@ -21,9 +21,7 @@ package metrics
 import (
 	"context"
 	"time"
-)
 
-import (
 	"dubbo.apache.org/dubbo-go/v3/common/constant"
 	"dubbo.apache.org/dubbo-go/v3/common/extension"
 	"dubbo.apache.org/dubbo-go/v3/filter"
@@ -52,7 +50,7 @@ func (p *Filter) Invoke(ctx context.Context, invoker protocol.Invoker, invocatio
 	duration := end.Sub(start)
 	go func() {
 		for _, reporter := range p.reporters {
-			reporter.Report(ctx, invoker, invocation, duration, res)
+			reporter.ReportAfterInvocation(ctx, invoker, invocation, duration, res)
 		}
 	}()
 	return res
